fixed GForceXY

+++ UDPF1Test/src/Motion.java
@@ -131,8 +131,8 @@ public class Motion {
         return Math.sqrt(m_gForceLateral*m_gForceLateral + m_gForceLongitudinal*m_gForceLongitudinal);
     }
 
-    public float getGForceXY(int decimals) {
-        return (float) Math.floor((Math.sqrt(m_gForceLateral * m_gForceLateral + m_gForceLongitudinal * m_gForceLongitudinal)*(Math.pow(10,decimals)))/Math.pow(10,decimals));
+    public double getGForceXY(int decimals) {
+        return Math.floor(Math.sqrt(m_gForceLateral * m_gForceLateral + m_gForceLongitudinal * m_gForceLongitudinal)*Math.pow(10,decimals))/Math.pow(10,decimals);
     }
 
     public double getGAngle(){
